+++ NEWS.md
@@ -1,7 +1,13 @@
 gen_socket
 ==========
 
-Changes with 0.3.0 - 19 Sep 2013
+Changes with 0.5.0 - 20 Jul 2015
+--------------------------------
+
+* add getsockopt/5 that passes an existing binary into the
+  getsockopt syscall
+
+Changes with 0.4.0 - 19 Sep 2013
 --------------------------------
 
 * reset the versioning to "Semantic Versioning"